refactor(move): refactor folder structure
