+++ apis/src/controllers/auth.controller.js
@@ -1,6 +1,6 @@
-const User = require("@schema/user.schema")
-const RefreshToken = require("@schema/refreshToken.schema")
-const SocialAccount = require("@schema/socialAccount.schema")
+const User = require("@models/User")
+const RefreshToken = require("@models/RefreshToken")
+const SocialAccount = require("@models/SocialAccount")
 const tokenUtils = require("@utils/token")
 const authUtils = require("@utils/auth")
 const emailUtils = require("@utils/email")
