+++ milestones.md
@@ -1,0 +1,24 @@
+# Key Milestones:
+
+- Week 1: Authentication and backend setup.
+- Week 2: Real-time chat functionality with Socket.IO.
+- Week 3: UI improvements, typing indicators, and user status.
+- Week 4: Group chat and media sharing.
+- Week 5: Testing, optimization, and error handling.
+- Week 6: User profile and settings.
+- Week 7: Final testing, security, and bug fixing.
+- Week 8: Documentation, deployment, and presentation.
+
+# Tracking
+
+1. **Week 1: Backend Setup and Authentication**
+
+- Implement user authentication:
+  - Set up user authentication with JWT (login and signup) if not already done.
+  - Securely hash passwords using bcrypt.
+  - Create protected routes on the backend.
+- Set up MongoDB models:
+  - Define the data structure for users, messages, and chat rooms.
+  - Create User, Message, and Room schemas using Mongoose.
+- Test Authentication:
+  - Test the login/signup process from the frontend and ensure token-based authentication is working.
